chatbot/dashboard/requirements.txt
Flask==2.3.3
pandas==2.1.4
Werkzeug==2.3.7
psycopg2-binary==2.9.7
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
